Migrate to RenderMarkdown from ArticleRenderer
