--- apps/trpfrog.net/src/app/blog/_renderer/MarkdownOptions.ts
@@ -1,5 +1,0 @@
-import { MDXRemoteProps } from 'next-mdx-remote/rsc'
-
-export type MarkdownOptions = Omit<MDXRemoteProps, 'source' | 'components'> & {
-  components: MDXRemoteProps['components']
-}
