--- apps/trpfrog.net/src/app/blog/_renderer/RenderInlineMarkdown.tsx
@@ -1,8 +1,0 @@
-import { MDXRemote } from 'next-mdx-remote/rsc'
-
-import { getMarkdownOptions } from '@blog/_renderer/rendererProperties.tsx'
-
-export const RenderInlineMarkdown = (props: { markdown: string }) => {
-  const options = getMarkdownOptions({ inline: true })
-  return <MDXRemote source={props.markdown} {...options} />
-}
