+++ apps/trpfrog.net/src/app/blog/_renderer/RenderMarkdown/types.ts
@@ -1,0 +1,5 @@
+import { BlogPost } from '@trpfrog.net/posts'
+
+export type MarkdownContext = Partial<{
+  blog: BlogPost
+}>
